+++ CHANGELOG.md
@@ -15,6 +15,13 @@ states, especially when defining custom functions.
     - [ProductStates::as_super_position] transforms the product state
       into a superposition with one amplitude.
 
+Fixes:
+
+- The labelling of the toffoli gate by the printer has changed from 'To'
+  to 'X'.
+- The labelling of the swap gate by the printer has changed from 'Swap'
+  to 'Sw'.
+
 ## 0.2.2 - Fixing the `Printer` and pushing of custom functions 
 
 Additions: